refactor(studio): split inline text style readers
